+++ R/general.R
@@ -1,4 +1,4 @@
-#' Check
+#' Check the dataset
 #'
 #' @param obj An object for check.
 #' @param dataset A dataset, a type is a data frame.
@@ -26,15 +26,30 @@ check.default <- function(obj, dataset) {
   # mutate empty strings in NA
   dataset <- dataset %>% dplyr::mutate(dplyr::across(.fns = ~ ifelse(. == "", NA, .)))
 
+  if (class(obj) == "lab") {
+    subjs <- dataset %>%
+      dplyr::filter(is.na(!!obj[["age"]]) | is.na(!!obj[["sex"]])) %>%
+      dplyr::select(!!obj[["id"]]) %>%
+      unlist() %>%
+      unique()
+
+    if (length(subjs) != 0) {
+      warning("problem with age or sex: ", paste(subjs, collapse = ","))
+    }
+  }
+
   # load file
   file <- obj[["file"]]
-  labs <- readxl::read_xlsx(file)
+  file_rows <- readxl::read_xlsx(file)
+  file_size <- nrow(file_rows)
+  pb <- progress::progress_bar$new(total = file_size)
 
   # create final dataset
   rs <- do.call(rbind, lapply(
-    seq_len(nrow(labs)),
+    seq_len(file_size),
     function(n) {
-      find_colnames(obj, dataset, labs[n, ])
+      pb$tick()
+      find_colnames(obj, dataset, file_rows[n, ])
     }
   ))
 
@@ -89,7 +104,7 @@ find_colnames.default <- function(obj, dataset, row_file) {
   do.call(
     rbind,
     lapply(parts, function(part) {
-      tryCatch(run_tests(obj, dataset, row_file, part),
+      tryCatch(to_long(obj, dataset, row_file, part),
         error = function(e) {
           warning(part, " can't bind, because ", e)
           data.frame()
@@ -99,7 +114,7 @@ find_colnames.default <- function(obj, dataset, row_file) {
   )
 }
 
-#' Get final result
+#' Get the final result of the check
 #'
 #' @param obj An object. Can be all classes: short, lab, date.
 #' @param group_id A logical scalar, default is TRUE.True is grouped by id, otherwise, it isn't grouped.
@@ -128,7 +143,7 @@ find_colnames.default <- function(obj, dataset, row_file) {
 #'
 #' obj_short <- check(obj_short, df)
 #' get_result(obj_short)
-get_result.default <- function(obj, group_id = T) {
+get_result <- function(obj, group_id = T) {
   result <- obj[["result"]]
 
   if (group_id) {
@@ -140,108 +155,6 @@ get_result.default <- function(obj, group_id = T) {
   result
 }
 
-#' Check sites
-#'
-#' @param objs A list of objects.
-#' @param dataset A dataset, a type is a data frame.
-#' @param col_site A column name of a site in the dataset, without quotes.
-#'
-#' @return A list of objects with a check result.
-#' @export
-#'
-#' @examples
-#' site <- c("site 01", "site 02")
-#' id <- c("01", "02")
-#' age <- c("19", "20")
-#' sex <- c("f", "m")
-#' gluc_post <- c("5.5", "4.1")
-#' gluc_res_post <- c("norm", "no")
-#' ast_post <- c("30", "48")
-#' ast_res_post <- c(NA, "norm")
-#'
-#' df <- data.frame(
-#'   site, id, age, sex,
-#'   gluc_post, gluc_res_post,
-#'   ast_post, ast_res_post,
-#'   stringsAsFactors = FALSE
-#' )
-#'
-#' refs_s01 <- system.file("labs_refer_s01.xlsx", package = "dmtools")
-#' refs_s02 <- system.file("labs_refer_s02.xlsx", package = "dmtools")
-#'
-#' s01_lab <- lab(refs_s01, id, age, sex, "norm", "no", site = "site 01")
-#' s02_lab <- lab(refs_s02, id, age, sex, "norm", "no", site = "site 02")
-#'
-#' labs <- list(s01_lab, s02_lab)
-#' labs <- check_sites(labs, df, site)
-check_sites <- function(objs, dataset, col_site) {
-  col_site <- dplyr::enquo(col_site)
-
-  objs %>% purrr::modify(function(obj) {
-    obj_site <- obj[["site"]]
-    pattern_site <- paste0("^", obj_site, "$")
-
-    # filter by site
-    data_site <- dataset %>%
-      dplyr::mutate(!!col_site := as.character(!!col_site)) %>%
-      dplyr::filter(grepl(pattern_site, !!col_site))
-
-    obj <- obj %>%
-      check(data_site)
-
-    obj[["result"]] <- obj[["result"]] %>% dplyr::mutate(num_site = obj_site)
-
-    obj
-  })
-}
-
-
-#' Test sites
-#'
-#' @param objs A list of objects.
-#' @param func A function e.g. \code{choose_test}, \code{get_result}.
-#'
-#' @return A data frame. The dataset.
-#' @export
-#'
-#' @examples
-#' site <- c("site 01", "site 02")
-#' id <- c("01", "02")
-#' age <- c("19", "20")
-#' sex <- c("f", "m")
-#' gluc_post <- c("5.5", "4.1")
-#' gluc_res_post <- c("norm", "no")
-#' ast_post <- c("30", "48")
-#' ast_res_post <- c(NA, "norm")
-#'
-#' df <- data.frame(
-#'   site, id, age, sex,
-#'   gluc_post, gluc_res_post,
-#'   ast_post, ast_res_post,
-#'   stringsAsFactors = FALSE
-#' )
-#'
-#' refs_s01 <- system.file("labs_refer_s01.xlsx", package = "dmtools")
-#' refs_s02 <- system.file("labs_refer_s02.xlsx", package = "dmtools")
-#'
-#' s01_lab <- lab(refs_s01, id, age, sex, "norm", "no", site = "site 01")
-#' s02_lab <- lab(refs_s02, id, age, sex, "norm", "no", site = "site 02")
-#'
-#' labs <- list(s01_lab, s02_lab)
-#' labs <- check_sites(labs, df, site)
-#'
-#' test_sites(labs, func = function(lab) choose_test(lab, "mis"))
-test_sites <- function(objs, func) {
-  result <- do.call(rbind, lapply(
-    objs,
-    function(obj) {
-      func(obj)
-    }
-  ))
-
-  row.names(result) <- seq_len(nrow(result))
-  result
-}
 
 #' Cast to double type
 #'